+++ contracts/modules/SSVClusters.sol
@@ -351,7 +351,7 @@ contract SSVClusters is ISSVClusters {
         address clusterOwner,
         uint64[] calldata operatorIds,
         Cluster memory cluster,
-        uint256 effectiveBalance,
+        uint32 effectiveBalance,
         bytes32[] calldata merkleProof
     ) external override {
         UpdateCtx memory ctx;
